Cleaning up a few things
